Installer setup Done 🧱

+++ includes/Frontend/Wac_auto.php
@@ -166,7 +166,7 @@ class Wac_auto
 											$discount = $wac_discounts["value"];
 											break;
 									}
-									$amount = (int) ($price - $discount);
+									$amount = (float) ($price - $discount);
 									$product->set_sale_price($amount);
 									return $amount;
 								}
@@ -181,7 +181,7 @@ class Wac_auto
 									$discount = $wac_discounts["value"];
 									break;
 							}
-							$amount = (int) ($price - $discount);
+							$amount = (float) ($price - $discount);
 							$product->set_sale_price($amount);
 							return $amount;
 						}
@@ -220,7 +220,7 @@ class Wac_auto
 										$discount = $wac_discounts["value"];
 										break;
 								}
-								$amount = (int) ($price - $discount);
+								$amount = (float) ($price - $discount);
 								$product->set_sale_price($amount);
 								return $amount;
 							}
@@ -235,7 +235,7 @@ class Wac_auto
 								$discount = $wac_discounts["value"];
 								break;
 						}
-						$amount = (int) ($price - $discount);
+						$amount = (float) ($price - $discount);
 						$product->set_sale_price($amount);
 						return $amount;
 					}
